Function rustms::algorithm::isotope::generate_isotope_distribution

pub fn generate_isotope_distribution(
    atomic_composition: &HashMap<String, i32>,
    mass_tolerance: f64,
    abundance_threshold: f64,
    max_result: i32,
) -> Vec<(f64, f64)>

generate the isotope distribution for a given atomic composition

Arguments:

atomic_composition - atomic composition of the peptide

mass_tolerance - mass tolerance for combining peaks

abundance_threshold - minimum abundance for a peak to be included in the result

max_result - maximum number of peaks to include in the result

Returns:

Vec<(f64, f64)> - distribution of masses and abundances

Examples

use std::collections::HashMap;
use rustms::algorithm::isotope::generate_isotope_distribution;

let mut atomic_composition = HashMap::new();
atomic_composition.insert("C".to_string(), 5);
atomic_composition.insert("H".to_string(), 9);
atomic_composition.insert("N".to_string(), 1);
atomic_composition.insert("O".to_string(), 1);
let result = generate_isotope_distribution(&atomic_composition, 1e-6, 1e-12, 200);